Pull the RelationMutationTests from cases/relation_test to cases/relation/mutation_test.
